Move shared code to shared module
