+++ esmvaltool/diag_scripts/land_carbon_cycle/diag_zonal_turnover.py
@@ -13,11 +13,11 @@ import matplotlib.pyplot as plt
 import numpy as np
 # import internal esmvaltool modules here
 from esmvaltool.diag_scripts.shared import group_metadata
-from esmvaltool.diag_scripts.shared import select_metadata
 from esmvaltool.diag_scripts.shared import run_diagnostic
 
 # helper functions
 import extraUtils as xu
+from .shared import _load_variable
 
 
 ## Classes and settings
@@ -71,17 +71,6 @@ def _apply_common_mask(dat_1, dat_2):
     dat_1 = np.ma.masked_invalid(dat_1)
     dat_2 = np.ma.masked_invalid(dat_2)
     return dat_1, dat_2
-
-
-def _apply_gpp_threshold(gpp_dat, fig_config):
-    '''
-    returns the input array with values below the threshold of gpp set to nan
-    '''
-    # converting gC m-2 yr-1 to kgC m-2 s-1
-    gpp_thres = fig_config.gpp_threshold / (86400.0 * 365 * 1000.)
-    gpp_dat = np.ma.masked_less(gpp_dat,
-                                gpp_thres).filled(fig_config.fill_value)
-    return gpp_dat
 
 
 def _get_obs_data(diag_config):
@@ -173,24 +162,6 @@ def _get_zonal_tau(diag_config):
     # plot the figure
     _plot_zonal_tau(zonal_tau_mod, zonal_tau_obs, diag_config)
     return 'done plotting the zonal turnover time'
-
-
-def _load_variable(metadata, var_name):
-    '''
-    load the data for a variable based on the metadata of the diagnostic
-    variable
-
-    Arguments:
-        metadata - nested dictionary of metadata
-
-    Returns:
-        iris cube of the data
-    '''
-    candidates = select_metadata(metadata, short_name=var_name)
-    assert len(candidates) == 1
-    filename = candidates[0]['filename']
-    cube = iris.load_cube(filename)
-    return cube
 
 
 def _calc_zonal_tau(dat_gpp, dat_ctotal, dat_lats, fig_config):
